pyproject.toml
[project]
name = "wabubu-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.52.0",
    "requests>=2.32.5",
]
